rename intention to testPurpose
