+++ src/gui/lib/operationHistory/actions/testPurpose/DeleteTestPurposeAction.ts
